Updated react-type-animation, added constants
