--- .eslintrc
@@ -1,7 +1,0 @@
-{
-	"extends": "react-app",
-	"plugins": ["prettier"],
-	"rules": {
-		"prettier/prettier": "error"
-	}
-}
