+++ README.md
@@ -8,14 +8,14 @@ My personal project portfolio developed by React
 
 [![react](https://img.shields.io/badge/react-16.10.2-green.svg)](https://github.com/PureDevPer/personal-website-react)
 [![react_router_dom](https://img.shields.io/badge/react_router_dom-5.1.2-green.svg)](https://github.com/PureDevPer/personal-website-react)
-[![react_type_animation](https://img.shields.io/badge/react_type_animation-1.6.2-green.svg)](https://github.com/PureDevPer/personal-website-react)
+[![react_type_animation](https://img.shields.io/badge/react_type_animation-3.1.0-green.svg)](https://github.com/PureDevPer/personal-website-react)
 [![react_helmet](https://img.shields.io/badge/react_helmet-5.2.1-green.svg)](https://github.com/PureDevPer/personal-website-react)
 [![styled_components](https://img.shields.io/badge/styled_components-4.4.0-green.svg)](https://github.com/PureDevPer/personal-website-react)
 [![fontawesome](https://img.shields.io/badge/fontawesome-5.11.2-green.svg)](https://github.com/PureDevPer/personal-website-react)
 
 I am a software engineer based in the United States.
 
-I'm interested in all aspects of software development. Recently, my passion lies in the web application development including both the front end and the back end. I fell in love with the modern JavaScript.
+I'm interested in all aspects of software development. Recently, my passion lies in the web application development including both the front end and the back end.
 
 Since I was a senior in college, I've liked mobile wireless networks as well. Therefore, I have read new telecommunication technologies such as next generation wireless networks.
 
@@ -23,21 +23,21 @@ Since I was a senior in college, I've liked mobile wireless networks as well. Th
 
 ### FULL STACK PROJECTS
 
-- MystaGram
-  - <a href="https://github.com/PureDevPer/Mystagram-FrontEnd">Front-End</a>
-  - <a href="https://github.com/PureDevPer/MystaGram">Back-End</a>
-  - <a href="https://github.com/PureDevPer/Mystagram-App">App</a>
-- <a href="https://github.com/PureDevPer/MyTube">MyTube</a>
-- <a href="https://github.com/PureDevPer/guessing_my_drawing">Guess My Drawing</a>
-- <a href="https://github.com/PureDevPer/BeerWorld">Beer World</a>
+-   MystaGram
+    -   <a href="https://github.com/PureDevPer/Mystagram-FrontEnd">Front-End</a>
+    -   <a href="https://github.com/PureDevPer/MystaGram">Back-End</a>
+    -   <a href="https://github.com/PureDevPer/Mystagram-App">App</a>
+-   <a href="https://github.com/PureDevPer/MyTube">MyTube</a>
+-   <a href="https://github.com/PureDevPer/guessing_my_drawing">Guess My Drawing</a>
+-   <a href="https://github.com/PureDevPer/BeerWorld">Beer World</a>
 
 ### WEB PROJECTS
 
-- <a href="https://github.com/PureDevPer/MovieWorld">Movie World</a>
-- <a href="https://github.com/PureDevPer/LearningJavaScript">Momentum</a>
-- <a href="https://github.com/PureDevPer/kakao">Kakao</a>
+-   <a href="https://github.com/PureDevPer/MovieWorld">Movie World</a>
+-   <a href="https://github.com/PureDevPer/LearningJavaScript">Momentum</a>
+-   <a href="https://github.com/PureDevPer/kakao">Kakao</a>
 
 ### OTHERS
 
-- <a href="https://github.com/PureDevPer/MachineLearning">Recognizing Hand-Written Characters</a>
-- <a href="https://www.youtube.com/watch?v=ijO5mvEtGKE&t=2s">Integrating Github with Jenkins</a>
+-   <a href="https://github.com/PureDevPer/MachineLearning">Recognizing Hand-Written Characters</a>
+-   <a href="https://www.youtube.com/watch?v=ijO5mvEtGKE&t=2s">Integrating Github with Jenkins</a>
