+++ package.json
@@ -1,48 +1,45 @@
 {
-	"name": "portfolio",
-	"version": "0.1.0",
-	"private": true,
-	"dependencies": {
-		"@fortawesome/fontawesome-svg-core": "^1.2.25",
-		"@fortawesome/free-brands-svg-icons": "^5.11.2",
-		"@fortawesome/free-regular-svg-icons": "^5.11.2",
-		"@fortawesome/free-solid-svg-icons": "^5.11.2",
-		"@fortawesome/react-fontawesome": "^0.1.7",
-		"gh-pages": "^2.1.1",
-		"prop-types": "^15.7.2",
-		"react": "^16.10.2",
-		"react-dom": "^16.10.2",
-		"react-helmet": "^5.2.1",
-		"react-router": "^5.1.2",
-		"react-router-dom": "^5.1.2",
-		"react-scripts": "3.2.0",
-		"react-typing-animation": "^1.6.2",
-		"react-youtube": "^7.9.0",
-		"styled-components": "^4.4.0",
-		"styled-reset": "^4.0.2"
-	},
-	"scripts": {
-		"start": "react-scripts start",
-		"build": "react-scripts build",
-		"test": "react-scripts test",
-		"eject": "react-scripts eject",
-		"deploy": "gh-pages -d build",
-		"predeploy": "npm run build"
-	},
-	"eslintConfig": {
-		"extends": "react-app"
-	},
-	"browserslist": {
-		"production": [
-			">0.2%",
-			"not dead",
-			"not op_mini all"
-		],
-		"development": [
-			"last 1 chrome version",
-			"last 1 firefox version",
-			"last 1 safari version"
-		]
-	},
-	"homepage": "https://puredevper.github.io/"
+    "name": "portfolio",
+    "version": "0.1.0",
+    "private": true,
+    "dependencies": {
+        "@fortawesome/fontawesome-svg-core": "^1.2.25",
+        "@fortawesome/free-brands-svg-icons": "^5.11.2",
+        "@fortawesome/free-regular-svg-icons": "^5.11.2",
+        "@fortawesome/free-solid-svg-icons": "^5.11.2",
+        "@fortawesome/react-fontawesome": "^0.1.7",
+        "gh-pages": "^2.1.1",
+        "prop-types": "^15.7.2",
+        "react": "^16.10.2",
+        "react-dom": "^16.10.2",
+        "react-helmet": "^5.2.1",
+        "react-router": "^5.1.2",
+        "react-router-dom": "^5.1.2",
+        "react-scripts": "^5.0.1",
+        "react-type-animation": "^3.1.0",
+        "react-youtube": "^7.9.0",
+        "styled-components": "^4.4.0",
+        "styled-reset": "^4.0.2"
+    },
+    "scripts": {
+        "start": "react-scripts start",
+        "build": "react-scripts build",
+        "test": "react-scripts test",
+        "eject": "react-scripts eject",
+        "deploy": "gh-pages -d build",
+        "predeploy": "npm run build"
+    },
+    "browserslist": {
+        "production": [
+            ">0.2%",
+            "not dead",
+            "not op_mini all"
+        ],
+        "development": [
+            "last 1 chrome version",
+            "last 1 firefox version",
+            "last 1 safari version"
+        ]
+    },
+    "homepage": "https://puredevper.github.io/"
 }
