+++ src/Components/constants/Introduction.js
@@ -1,0 +1,21 @@
+export const INTRODUCTION = {
+    INTRO: `
+        Hi! I'm Wooseok.
+        I'm a Software Engineer based in the United States.
+        I'm currently located in the Bay Area, California.`,
+    INTERESTS: `
+        I'm interested in all aspects of software development.
+        Recently, my passion lies in the web application development including both the front end and the back end.`,
+    INTERESTS_OTHER: `
+        Since I was a senior in college, I've liked mobile wireless networks as well.
+        Therefore, I have read new telecommunication technologies such as next generation wireless networks.`,
+    HOBBY: `
+            Outside the realm of computers, I enjoy car detailing as a car enthusiast.
+            I research how to make the car cleaner.
+            I also like traveling all over the world.
+            One of my dreams is to visit all the national parks in the United States.`,
+    PREV: `
+        Previously, I taught code to undergraduate students in college,
+        worked as a software engineer for LG and developed web applications as a software engineer intern.
+        I also hold a Bachelor's and Master's Degree in computer science.`,
+};
