+++ src/Components/constants/Landing.js
@@ -1,0 +1,6 @@
+export const WOOSEOK_KIM = "Wooseok Kim";
+
+export const TYPE_ANIMATION = {
+    SEQUENCE: ["Software", 500, "Software Engineer", 800, "Software", 500, "", 800],
+    STYLE: { fontWeight: 600, fontSize: "25px", color: "white" },
+};
